Update PlaceHarmonizer.js
Added code to handle settings checkboxes, including handling of saving of settings to local storage.  Currently, the settings themselves do not actually do anything.  This needs to be handled yet.

TODO:  The checkbox logic must still be handled so that they actually do something in the rest of the code.